More bindSchedulingOperators to a separate file
